Showing delete column only if user is logged in and its admin

+++ src/components/moviesTable.jsx
@@ -2,6 +2,7 @@ import React, { Component } from "react";
 import Like from "./common/like";
 import Table from "./common/table";
 import { Link } from "react-router-dom";
+import authService from "../services/authService";
 
 class MoviesTable extends Component {
   columns = [
@@ -18,19 +19,30 @@ class MoviesTable extends Component {
       content: movie => (
         <Like liked={movie.liked} onClick={() => this.props.onLike(movie)} />
       )
-    },
-    {
-      key: "delete",
-      content: movie => (
-        <button
-          onClick={() => this.props.onDelete(movie._id)}
-          className="btn btn-danger btn-sm"
-        >
-          Delete
-        </button>
-      )
     }
   ];
+
+  constructor() {
+    super();
+    const user = authService.getCurrentUser();
+
+    if (user && user.isAdmin) {
+      // if the user is logged-in and its admin
+      // then we add delete column to the columns
+
+      this.columns.push({
+        key: "delete",
+        content: movie => (
+          <button
+            onClick={() => this.props.onDelete(movie._id)}
+            className="btn btn-danger btn-sm"
+          >
+            Delete
+          </button>
+        )
+      });
+    }
+  }
 
   render() {
     const { movies, sortColumn, onSort } = this.props;
